controller -service -part 3

See merge request incubateur_m2i_afpa_2/team-ecom/back-end!7

+++ src/main/java/fr/insy2s/commerce/dtos/AdresseRequest.java
@@ -7,9 +7,9 @@ import lombok.Setter;
 @Setter
 public class AdresseRequest {
 
-    private int adresseId;
+    private Long adresseId;
 
-    private int userId;
+    private Long userId;
 
 
 }
